feat(listing): new grants ui  with layout
