+++ src/features/grants/components/GrantsHeader.tsx
@@ -13,7 +13,12 @@ import React from 'react';
 import { LuCheck } from 'react-icons/lu';
 
 import { VerifiedBadge } from '@/components/shared/VerifiedBadge';
-import { ListingTabLink, RegionLabel, StatusBadge } from '@/features/listings';
+import {
+  ListingHeaderSeparator,
+  ListingTabLink,
+  RegionLabel,
+  StatusBadge,
+} from '@/features/listings';
 import type { SponsorType } from '@/interface/sponsor';
 import { PulseIcon } from '@/svg/pulse-icon';
 
@@ -97,12 +102,6 @@ export const GrantsHeader = ({
                 >
                   {title}
                 </Heading>
-
-                <StatusBadge
-                  Icon={statusIcon}
-                  textColor={statusTextColor}
-                  text={statusText}
-                />
               </Flex>
 
               <Flex align={'center'} wrap={'wrap'} gap={{ base: 1, md: 3 }}>
@@ -117,21 +116,36 @@ export const GrantsHeader = ({
                   </Text>
                   {!!sponsor?.isVerified && <VerifiedBadge />}
                 </Flex>
-                <Text color={'#E2E8EF'} fontWeight={500}>
-                  |
-                </Text>
+                <ListingHeaderSeparator />
+                <Flex>
+                  <Image
+                    h="4"
+                    mt={{ base: '1px', sm: 1 }}
+                    mr={{ base: '1px', sm: 1 }}
+                    alt={'grant'}
+                    src={'/assets/icons/bank.svg'}
+                  />
+                  <Text
+                    color={'brand.slate.400'}
+                    fontSize={{ base: 'xs', sm: 'md' }}
+                    fontWeight={500}
+                  >
+                    Grant
+                  </Text>
+                </Flex>
+                <ListingHeaderSeparator />
+                <StatusBadge
+                  Icon={statusIcon}
+                  textColor={statusTextColor}
+                  text={statusText}
+                />
+                <ListingHeaderSeparator />
                 <RegionLabel region={region} />
               </Flex>
             </VStack>
           </HStack>
         </VStack>
-        <Flex
-          align={'center'}
-          w={'full'}
-          h={10}
-          borderTop={'1px solid'}
-          borderTopColor={'gray.100'}
-        >
+        <Flex align={'center'} w={'full'} h={10}>
           <HStack
             align="center"
             justifyContent="start"
@@ -142,18 +156,30 @@ export const GrantsHeader = ({
             mx={'auto'}
             my={'auto'}
             px={3}
+            borderColor="brand.slate.200"
+            borderBottomWidth={'1px'}
           >
             <ListingTabLink
+              w={{ md: '22rem' }}
               href={`/grants/${slug}/`}
-              text="DETAILS"
-              isActive={!router.asPath.includes('references')}
+              text="Prizes"
+              isActive={false}
+              styles={{
+                pointerEvents: 'none',
+                display: { base: 'none', md: 'flex' },
+              }}
+            />
+            <ListingTabLink
+              href={`/grants/${slug}/`}
+              text="Details"
+              isActive={!router.asPath.split('/')[3]?.includes('references')}
             />
 
             {references && references?.length > 0 && (
               <ListingTabLink
                 href={`/grants/${slug}/references`}
-                text="REFERENCES"
-                isActive={router.asPath.includes('references')}
+                text="References"
+                isActive={!!router.asPath.split('/')[3]?.includes('references')}
               />
             )}
           </HStack>
