+++ src/features/listings/components/ListingPage/Separator.tsx
@@ -1,0 +1,9 @@
+import { Text } from '@chakra-ui/react';
+
+export const ListingHeaderSeparator = () => {
+  return (
+    <Text color={'#E2E8EF'} fontWeight={500}>
+      |
+    </Text>
+  );
+};
