+++ src/features/listings/components/ListingPage/index.ts
@@ -1,8 +1,10 @@
 export * from './DescriptionUI';
+export * from './ExtraInfoSection';
 export * from './ListingHeader';
 export * from './ListingTabLink';
 export * from './ListingWinners';
 export * from './ReferenceCard';
 export * from './RegionLabel';
 export * from './RightSideBar';
+export * from './Separator';
 export * from './StatusBadge';
