+++ src/features/listings/types/index.ts
@@ -44,16 +44,7 @@ export interface Listing {
   referredBy?: SuperteamName;
   publishedAt?: string;
   isPrivate?: boolean;
-  Hackathon?: {
-    name: string;
-    logo: string;
-    description: string;
-    deadline: string;
-    startDate: string;
-    altLogo: string;
-    slug: string;
-    announceDate: string;
-  };
+  Hackathon?: ListingHackathon;
   compensationType?: 'fixed' | 'range' | 'variable';
   minRewardAsk?: number;
   maxRewardAsk?: number;
@@ -61,6 +52,17 @@ export interface Listing {
   _count?: {
     Comments?: number;
   };
+}
+
+export interface ListingHackathon {
+  name: string;
+  logo: string;
+  description: string;
+  deadline: string;
+  startDate: string;
+  altLogo: string;
+  slug: string;
+  announceDate: string;
 }
 
 export interface ListingWithSubmissions extends Listing {
